good practices for alarm linux
