+++ lib/agent/actions/alarm/index.js
@@ -13,8 +13,8 @@ var fs       = require('fs'),
     os_name  = process.platform.replace('darwin', 'mac').replace('win32', 'windows'),
     commands = require('./' + os_name),
     Emitter  = require('events').EventEmitter,
-    system   = require('./../../common').system,
-    spawn = require('child_process').spawn,
+    system   = require('./../../common').system, 
+    linux_alarm = require('./linux'),
     is_linux = os_name == 'linux';
 
 var child,
@@ -40,64 +40,12 @@ exports.start = function(options, cb) {
     emitter = null;
   }
 
-  var setVolume = function (val, cb) {
-    defaultDevice(function (err, dev) {
-      if (err) {
-      } else {
-        amixer(['set', dev, val + '%'], function (err) {
-        });
-      }
-    });
-  };
-
   var raise_volume = function(cb) {
-    if (is_linux) setVolume(volumen)
+    if (is_linux) linux_alarm.setVolume(volumen);
     else
       exec(commands.raise_volume, cb)
   }
 
-  var amixer = function (args, cb) {
-
-    var ret = '';
-    var err = null;
-    console.log(args)
-    var p = spawn('amixer', args);
-
-    p.stdout.on('data', function (data) {
-      ret += data;
-    });
-
-    p.stderr.on('data', function (data) {
-      err = new Error('Alsa Mixer Error: ' + data);
-    });
-
-    p.on('close', function () {
-      cb(err, ret.trim());
-    });
-
-  };
-
-  var reDefaultDevice = /Simple mixer control \'([a-z0-9 -]+)\',[0-9]+/i;
-  var defaultDeviceCache = null;
-  var defaultDevice = function (cb) {
-    if (defaultDeviceCache === null) {
-      amixer([], function (err, data) {
-        if (err) {
-          cb(err);
-        } else {
-          var res = reDefaultDevice.exec(data);
-          if (res === null) {
-            cb(new Error('Alsa Mixer Error: failed to parse output'));
-          } else {
-            defaultDeviceCache = res[1];
-            cb(null, defaultDeviceCache);
-          }
-        }
-      });
-    } else {
-      cb(null, defaultDeviceCache);
-    }
-  };
 
 
   var play_queue = function() {
