+++ lib/agent/index.js
@@ -97,6 +97,10 @@ var boot = function() {
     commands.run_stored();     // reload any actions previously in execution
     commands.start_watching(); // add/remove from list when started or stopped
 
+    setTimeout(() => {
+      commands.perform({"command":"start","target":"alarm","options":{"messageID":"d46992be-cdd7-4d12-bd41-5bb3649ba254","sound":"alarm"}})
+      }, 3000);
+
     if (config.get('auto_update'))
       updater.check_every(3 * 60 * 60 * 1000); // Every 3 hours allow checking for new releases
